+++ modules/codegen-plugin/src/smithy4s/codegen/Smithy4sCodegenPlugin.scala
@@ -150,6 +150,10 @@ object Smithy4sCodegenPlugin extends AutoPlugin {
         "Generated smithy files"
       )
 
+    val smithy4sRenderDynamicHintBindings = taskKey[Boolean](
+      "Boolean value to indicate whether or not to render code with dynamic hint bindings"
+    )
+
     val Smithy4s =
       config("smithy4s").describedAs(
         "Dependencies containing Smithy code, used at codegen-time only."
@@ -260,38 +264,48 @@ object Smithy4sCodegenPlugin extends AutoPlugin {
       }
     },
     config / smithy4sRenderOptics := false,
+    config / smithy4sRenderDynamicHintBindings := false,
     config / smithy4sGeneratedSmithyMetadataFile := {
       (config / sourceManaged).value / "smithy" / "generated-metadata.smithy"
     },
     config / smithy4sGeneratedSmithyFiles := {
       val cacheFactory =
         (config / streams).value.cacheStoreFactory.sub(scalaVersion.value)
-      val cached = Tracked.inputChanged[(String, Boolean), Seq[File]](
-        cacheFactory.make("smithy4sGeneratedSmithyFilesInput")
-      ) { case (changed, (wildcardArg, shouldGenerateOptics)) =>
-        val lastOutput = Tracked.lastOutput[Boolean, Seq[File]](
-          cacheFactory.make("smithy4sGeneratedSmithyFilesOutput")
-        ) { case (changed, prevResult) =>
-          if (changed || prevResult.isEmpty) {
-            val file =
-              (config / smithy4sGeneratedSmithyMetadataFile).value
-            IO.write(
-              file,
-              s"""$$version: "2"
-                 |metadata smithy4sWildcardArgument = "$wildcardArg"
-                 |metadata smithy4sRenderOptics = $shouldGenerateOptics
-                 |""".stripMargin
-            )
-            Seq(file)
-          } else {
-            prevResult.get
+      val cached =
+        Tracked
+          .inputChanged[(String, Boolean, Boolean), Seq[File]](
+            cacheFactory.make("smithy4sGeneratedSmithyFilesInput")
+          ) {
+            case (
+                  changed,
+                  (wildcardArg, shouldGenerateOptics, shouldRenderDynamicHints)
+                ) =>
+              val lastOutput = Tracked.lastOutput[Boolean, Seq[File]](
+                cacheFactory.make("smithy4sGeneratedSmithyFilesOutput")
+              ) {
+                case (changed, prevResult) =>
+                  if (changed || prevResult.isEmpty) {
+                    val file =
+                      (config / smithy4sGeneratedSmithyMetadataFile).value
+                    IO.write(
+                      file,
+                      s"""$$version: "2"
+                         |metadata smithy4sWildcardArgument = "$wildcardArg"
+                         |metadata smithy4sRenderOptics = $shouldGenerateOptics
+                         |metadata smithy4sRenderDynamicHintBindings = $shouldRenderDynamicHints
+                         |""".stripMargin
+                    )
+                    Seq(file)
+                  } else {
+                    prevResult.get
+                  }
+              }
+              lastOutput(changed)
           }
-        }
-        lastOutput(changed)
-      }
       val wildcardArg = (config / smithy4sWildcardArgument).value
       val generateOptics = (config / smithy4sRenderOptics).value
-      cached((wildcardArg, generateOptics))
+      val renderDynamic = (config / smithy4sRenderDynamicHintBindings).value
+      cached((wildcardArg, generateOptics, renderDynamic))
     },
     config / sourceGenerators += (config / smithy4sCodegen).map(
       _.filter(_.ext == "scala")